requirements.txt
fastapi
uvicorn
yfinance
matplotlib
python-dotenv
requests
pmdarima
keras
tensorflow
prophet

